Renames Foundatio.Xunit2 to Foundatio.Xunit
Renames the Foundatio.Xunit2 project to Foundatio.Xunit to align with xUnit v2.
Updates project references and namespaces accordingly.

This change prepares for removing xUnit v2 support and
consolidating the xUnit helpers into a single project.
