+++ Foundatio.All.slnx
@@ -56,7 +56,7 @@
   <Project Path="src/Foundatio.JsonNet/Foundatio.JsonNet.csproj" />
   <Project Path="src/Foundatio.MessagePack/Foundatio.MessagePack.csproj" />
   <Project Path="src/Foundatio.Utf8Json/Foundatio.Utf8Json.csproj" />
-  <Project Path="src/Foundatio.Xunit2/Foundatio.Xunit2.csproj" />
+  <Project Path="src/Foundatio.Xunit/Foundatio.Xunit.csproj" />
   <Project Path="src/Foundatio.Xunit.v3/Foundatio.Xunit.v3.csproj" />
   <Project Path="src/Foundatio/Foundatio.csproj" />
 </Solution>
