+++ src/Foundatio.Xunit/Logging/TestLoggerBase.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Foundatio.Xunit2;
+namespace Foundatio.Xunit;
 
 public abstract class TestLoggerBase : IClassFixture<TestLoggerFixture>, IAsyncLifetime
 {
