+++ src/Foundatio.Xunit/Logging/TestLoggerOptions.cs
@@ -4,7 +4,7 @@ using Foundatio.Utility;
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;
 
-namespace Foundatio.Xunit2;
+namespace Foundatio.Xunit;
 
 public class TestLoggerOptions
 {
