+++ src/Foundatio.Xunit/Logging/TestWithLoggingBase.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;
 
-namespace Foundatio.Xunit2;
+namespace Foundatio.Xunit;
 
 public abstract class TestWithLoggingBase
 {
